Update: Move schedule app urls, forms and views to xero_workspace app - delete schedule app and comment its migration dependencies in xero_workspace app migrations - remove schedule app from installed apps in settings file
