+++ apps/xero_workspace/templates/xero_workspace/schedule.html
@@ -4,7 +4,7 @@
 
 {% block link %}
     <link rel="stylesheet" href="{% static 'css/navtab.css' %}">
-    <link rel="stylesheet" href="{% static 'css/schedule/schedule.css' %}">
+    <link rel="stylesheet" href="{% static 'css/xero_workspace/schedule.css' %}">
     {{ form.media }}
 {% endblock %}
