+++ fyle_xero_integration_web_app/settings.py
@@ -46,7 +46,6 @@ INSTALLED_APPS = [
     'apps.user',
     'apps.fyle_connect',
     'apps.xero_workspace',
-    'apps.schedule',
     'apps.sync_activity',
     'tempus_dominus',
